requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==0.10.0
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.21
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
future==0.18.2
google-auth==1.21.0
google-auth-oauthlib==0.4.1
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
kiwisolver
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.1
mccabe==0.6.1
msgpack==1.0.0
neovim==0.3.1
oauthlib==3.1.0
pandas==1.0.5
parso==0.7.0
passlib==1.7.2
Pillow==7.2.0
prettytable==1.0.0
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pynvim==0.4.1
pyparsing==2.4.7
pytesseract==0.3.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy
six==1.15.0
SQLAlchemy==1.3.19
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
toml==0.10.1
torch
torchvision
tqdm==4.48.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
opencv-python
